requirements.txt
django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.7,<4.0
psycopg2-binary>=2.9,<3.0
django-environ>=0.4,<1.0
django-filter>=2.4,<3.0
Pillow>=9.0,<10.0
requests>=2.25,<3.0
pytest>=6.0,<7.0
black>=21.0,<22.0
flake8>=3.9,<4.0
django-extensions>=3.1,<4.0
drf-yasg>=1.21,<2.0
